action/requirements.txt
request>=2.28.1
PyGithub==1.55